backend/requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
psycopg2-binary==2.9.10
requests==2.32.5
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
django-extensions>=4.1
django-cors-headers>=4.8.0
